Testing/backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
python-decouple==3.8
celery==5.3.6
redis==5.0.1
pillow==10.2.0
drf-yasg==1.21.7
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
